agentic-knowledge-rag-graph/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
psycopg2-binary==2.9.9
qdrant-client==1.7.0
ollama==0.2.1
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
networkx==3.2.1
sentence-transformers==2.2.2
langchain==0.0.348
langchain-community==0.0.1